+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- The `Drop` impl now uses the `mem::needs_drop()` optimization hint to avoid
+  unnecessary overhead.
 
 ## [0.2.0] - 2017-09-17
 ### Added